[refactor] get room response by room entity
